Vastly improve wall kick implementation (not fully complete)

+++ include/Tetronimo.h
@@ -84,7 +84,7 @@ constexpr Color PIECE_COLOR[] = {
 
 struct Tetronimo {
   Piece    ty;
-  usize    rotation;
+  u8       rotation;
   Position pos;
 };
